fix(shopwired): use bigInteger for webhook_id and cross-reference duplicate WebhookTopic enums
webhook_id changed from integer (max 2.1B) to bigInteger to prevent
future overflow as ShopWired IDs grow. Both WebhookTopic enums now
cross-reference each other with IMPORTANT flags to prevent drift.

+++ database/migrations/2026_03_15_100000_create_shopwired_webhook_events_table.php
@@ -22,7 +22,7 @@ return new class extends Migration {
 
             $table->integer('subject_id');
             $table->string('topic', 50);
-            $table->integer('webhook_id')->unique();
+            $table->bigInteger('webhook_id')->unique();
             $table->timestampTz('event_time');
 
             $table->timestampTz('created_at');
